+++ spec/transpec/commit_message_spec.rb
@@ -11,7 +11,7 @@ module Transpec
     subject(:commit_message) { CommitMessage.new(report, rspec_version, cli_args) }
     let(:report) { Report.new }
     let(:rspec_version) { RSpecVersion.new('2.99.0.beta1') }
-    let(:cli_args) { %w(--force --generate-commit-message) }
+    let(:cli_args) { ['--force'] }
 
     before do
       report.records << Record.new('obj.stub(:message)', 'allow(obj).to receive(:message)')
@@ -47,7 +47,7 @@ module Transpec
           body_lines[0].chomp
             .should match(/^This conversion is done by Transpec \d+\.\d+\.\d+ with the following command:$/)
           body_lines[1].chomp
-            .should ==     '    transpec --force --generate-commit-message'
+            .should ==     '    transpec --force'
         end
 
         it 'has blank line after the preface' do