+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Development
 
+## v1.5.0
+
 * Mention the project's RSpec version in the auto-generated commit message
 * Deprecate `-m/--generate-commit-message` option (A commit message will always be generated)
 * Display actual `Proc` constructor names in the summary of expectations on block